// backend/src/features/session/connect/schema.ts
import { FastifySchema, RequestGenericInterface } from "fastify";

import { Static, Type } from "@sinclair/typebox";

const Body = Type.Object({
  key: Type.String({
    description: "The key of the generated code.",
  }),
  device: Type.Optional(
    Type.Object({
      ip: Type.Optional(Type.String({ description: "The ip of the device" })),
      agent: Type.Optional(
        Type.String({ description: "The agent attached to the device" })
      ),
    })
  ),
});

const Reply = Type.Object({
  token: Type.String({ description: "The client token." }),
  device: Type.Number({ description: "The id of the device" }),
  session: Type.Number({ description: "The id of the session" }),
});

export interface Interface extends RequestGenericInterface {
  Body: Static<typeof Body>;
  Reply: Static<typeof Reply>;
}

export const Schema: FastifySchema = {
  tags: ["session"],
  description:
    "Generate a token for a session from a code generated by the host. Use this token to call push messages to the host.",
  body: Body,
  response: {
    200: Reply,
  },
};